+++ examples/example/vue3/src/examples/auto-complete/01/App.vue
@@ -1,8 +1,9 @@
 <script setup lang="ts">
 import { ref } from 'vue'
 import { AutoComplete } from '@expcat/tigercat-vue/AutoComplete'
+import { FormItem } from '@expcat/tigercat-vue/FormItem'
 
-const value = ref('')
+const value = ref<string | number | undefined>()
 const options = ['React', 'Vue', 'Svelte', 'Angular'].map((item) => ({
   label: item,
   value: item
@@ -10,9 +11,7 @@ const options = ['React', 'Vue', 'Svelte', 'Angular'].map((item) => ({
 </script>
 
 <template>
-  <AutoComplete
-    v-model="value"
-    :options="options"
-    placeholder="输入框架名称"
-    class="w-full max-w-sm" />
+  <FormItem label="框架" class="w-full max-w-sm">
+    <AutoComplete v-model="value" :options="options" placeholder="输入框架名称" clearable />
+  </FormItem>
 </template>